requirements.txt
langchain
langchain-groq
tavily-python
python-dotenv
flask
sentence-transformers
faiss-cpu


